Relocate schema folder inside model folder
