+++ src/post/model/Post.ts
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import postSchema from "../controller/PostsController/schema/postSchema.js";
+import postSchema from "./schema/postSchema.js";
 
 const Post = mongoose.model("Posts", postSchema, "posts");
